python-backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
aiomysql
redis
python-dotenv
pydantic-settings
httpx
stripe
greenlet
langchain
langchain-openai
